ios version check, support for cupertino-decompose
