+++ cupertino/src/commonMain/kotlin/io/github/alexzhirkevich/cupertino/theme/CupertinoTheme.kt
@@ -43,6 +43,7 @@ import androidx.compose.ui.platform.LocalHapticFeedback
 import io.github.alexzhirkevich.LocalContentColor
 import io.github.alexzhirkevich.LocalTextStyle
 import io.github.alexzhirkevich.cupertino.ExperimentalCupertinoApi
+import io.github.alexzhirkevich.cupertino.SystemBarAppearance
 import io.github.alexzhirkevich.cupertino.rememberCupertinoHapticFeedback
 import io.github.alexzhirkevich.cupertino.rememberCupertinoIndication
