--- cupertino/src/commonMain/kotlin/io/github/alexzhirkevich/cupertino/theme/SystemBarAppearance.kt
@@ -1,6 +1,0 @@
-package io.github.alexzhirkevich.cupertino.theme
-
-import androidx.compose.runtime.Composable
-
-@Composable
-internal expect fun SystemBarAppearance(dark: Boolean)
